Remove the eslint disable line and fix the issue in the code.
